Extract use cases from ViewModel
