+++ features/dashboard/ui/src/main/kotlin/com/boswelja/truemanager/dashboard/ui/DI.kt
@@ -1,6 +1,6 @@
 package com.boswelja.truemanager.dashboard.ui
 
-import com.boswelja.truemanager.dashboard.ui.OverviewViewModel
+import com.boswelja.truemanager.dashboard.ui.overview.OverviewViewModel
 import org.koin.androidx.viewmodel.dsl.viewModelOf
 import org.koin.dsl.module
