+++ features/dashboard/ui/src/main/kotlin/com/boswelja/truemanager/dashboard/ui/overview/OverviewScreen.kt
@@ -10,7 +10,6 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.boswelja.truemanager.dashboard.ui.OverviewViewModel
 import com.boswelja.truemanager.dashboard.ui.overview.cards.CpuCard
 import com.boswelja.truemanager.dashboard.ui.overview.cards.MemoryCard
 import com.boswelja.truemanager.dashboard.ui.overview.cards.NetworkCard
